requirements.txt
asgiref==3.3.4
boto3==1.17.69
botocore==1.20.69
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.25.0
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-storages==1.11.1
django-summernote==0.8.11.6
django-taggit==1.4.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Pillow==8.2.0
ply==3.11
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
